Rework interface for reactive 2.0
This includes a major change to the state modelling; specifically
'endpoint.{endpoint_name}.new-request' is set whenever a relation
data change event is detected for any of the supported keys on
the relation.

The consuming charm must clear this event at the end of its handler
invocation for 'endpoint.{endpoint_name}.new-request'.